Moving all terminal logic to human.py
