+++ pyminion/expansions/base.py
@@ -1,12 +1,10 @@
 import logging
 import math
-from typing import TYPE_CHECKING, List, Optional, Tuple, Union
+from typing import TYPE_CHECKING, List, Tuple, Union
 
 from pyminion.bots.bot import Bot
 from pyminion.core import AbstractDeck, CardType, Action, Card, Treasure, Victory
-from pyminion.decisions import validate_input
-from pyminion.exceptions import (EmptyPile, InvalidBotImplementation,
-                                 InvalidMultiCardInput, InvalidSingleCardInput)
+from pyminion.exceptions import EmptyPile
 from pyminion.human import Human
 from pyminion.player import Player
